fix: use existing BRP fixture file
Merge new fixtures with existing BRP ones

Solves: PZ-4479
